Q&A_CHATBOT/requirements.txt
langchain-ollama
langchain
python-dotenv
langchain_core
streamlit
langchain-groq 
gtts 
speechrecognition
pyaudio
PyMuPDF
chromadb
sentence-transformers
ipykernel
langchain-community
pypdf
bs4
arxiv
wikipedia
langchain-text-splitters
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
pdfplumber